refactor: 유틸함수 역할끼리 파일 분리
