+++ frontend/src/components/Calendar/Calendar.tsx
@@ -11,14 +11,14 @@ import useCalendar from '~/hooks/useCalendar';
 import { useScheduleModal } from '~/hooks/schedule/useScheduleModal';
 import { useFetchSchedules } from '~/hooks/queries/useFetchSchedules';
 import { useModal } from '~/hooks/useModal';
-import { generateScheduleBars } from '~/utils/generateScheduleBars';
+import { generateScheduleBars } from '~/utils/generate/generateScheduleBars';
 import { DAYS_OF_WEEK, MODAL_OPEN_TYPE } from '~/constants/calendar';
 import { ArrowLeftIcon, ArrowRightIcon, PlusIcon } from '~/assets/svg';
 import { arrayOf } from '~/utils/arrayOf';
 import ScheduleMoreCell from '~/components/ScheduleMoreCell/ScheduleMoreCell';
 import type { Position, ModalOpenType } from '~/types/schedule';
 import DailyScheduleModal from '~/components/DailyScheduleModal/DailyScheduleModal';
-import { getDateByPosition } from '~/utils/getDateByPosition';
+import { getDateByPosition } from '~/utils/get/getDateByPosition';
 import { useTeamPlace } from '~/hooks/useTeamPlace';
 
 const Calendar = () => {
