+++ frontend/src/components/MyDailyScheduleList/MyDailySchedule/MyDailySchedule.tsx
@@ -4,7 +4,7 @@ import type { TeamPlaceColor } from '~/types/team';
 import Text from '~/components/common/Text/Text';
 import { ClockIcon } from '~/assets/svg';
 import { GroupIcon } from '~/assets/svg';
-import { formatTime } from '~/utils/formatTime';
+import { formatTime } from '~/utils/format/formatTime';
 
 export interface MyDailyScheduleProps extends ScheduleWithoutId {
   teamPlaceColor: TeamPlaceColor;
