+++ frontend/src/contexts/TeamPlaceContext.tsx
@@ -2,7 +2,7 @@ import { createContext, useCallback, useEffect, useState } from 'react';
 import type { PropsWithChildren } from 'react';
 import { useFetchTeamPlaces } from '~/hooks/queries/useFetchTeamPlaces';
 import type { TeamPlace, TeamPlaceColor } from '~/types/team';
-import { getInfoByTeamPlaceId } from '~/utils/getInfoByTeamPlaceId';
+import { getInfoByTeamPlaceId } from '~/utils/get/getInfoByTeamPlaceId';
 
 interface TeamPlaceContextProps {
   teamPlaces: TeamPlace[];
