+++ frontend/src/utils/format/formatISOString.test.ts
@@ -1,4 +1,4 @@
-import { formatISOString } from '~/utils/formatISOString';
+import { formatISOString } from '~/utils/format/formatISOString';
 
 describe('formatISOString', () => {
   it('YYYY-MM-DDTHH:MM 형식의 ISO string을 YYYY-MM-DD HH:MM 형식으로 변환한다.', () => {
