+++ frontend/src/utils/get/getInfoByTeamPlaceId.ts
